Update test and linters
