+++ .rubocop.yml
@@ -26,7 +26,7 @@ Metrics/ClassLength:
   Max: 150
 Metrics/BlockLength:
   AllowedMethods: ['describe']
-  Max: 30
+  Max: 40
 
 Style/Documentation:
   Enabled: false
