+++ app/controllers/posts_controller.rb
@@ -2,7 +2,6 @@ class PostsController < ApplicationController
   def index
     @user = User.find(params[:user_id])
     @posts = @user.posts.includes(:comments)
-    
   end
 
   def show
