+++ app/views/posts/index.html.erb
@@ -1,7 +1,7 @@
 <%= stylesheet_link_tag 'style' %>
 
 <div class="container">
-<img src ='<%= @user.photo %>' class="userImg" alt="User Photo">
+<img src ='<%= @user.photo %>'  class="userImg" alt="User Photo">
 <div class="userCon">
 <h1 class="userName"><%= @user.name %> </h1>
 <p class="Upostcount">Number of Posts: <%= @user.posts.count %></p>
