+++ app/views/users/index.html.erb
@@ -2,8 +2,8 @@
 
 <div >
   <% @users.each do |user| %>
-    <a class="container"  href="<%= user_path(user.id) %>" class="userLink">
-      <img src ='<%= user.photo %>' class="userImg" alt="User Photo">
+    <a class="container"  href="<%= user_path(user.id) %>" >
+      <img src ='<%= user.photo %>'  class="userImg" alt="User Photo">
       <div class="userCon">
         <h1 class="userName"><%= user.name %></h1>
         <p class="Upostcount">Number of posts: <%= user.posts.count %></p>
